lambda/requirements.txt
ask-sdk-core
ask-sdk-s3-persistence-adapter
beautifulsoup4
random-word